Create separate feature for adyen and for trustly
