+++ app/feature/feature-connect-payment-adyen/src/main/AndroidManifest.xml
