+++ app/feature/feature-connect-payment-adyen/src/main/kotlin/com/hedvig/android/feature/connect/payment/adyen/ConnectAdyenPaymentGraph.kt
@@ -1,12 +1,10 @@
-package com.hedvig.android.feature.connect.payment
+package com.hedvig.android.feature.connect.payment.adyen
 
 import androidx.navigation.NavGraphBuilder
 import com.hedvig.android.navigation.core.AppDestination
 import com.kiwi.navigationcompose.typed.composable
 
-fun NavGraphBuilder.connectPaymentGraph() {
-  composable<AppDestination.ConnectPaymentTrustly>() {
-  }
+fun NavGraphBuilder.connectAdyenPaymentGraph() {
   composable<AppDestination.ConnectPaymentAdyen>() {
   }
 }
