+++ app/feature/feature-connect-payment-trustly/src/main/kotlin/com/hedvig/android/feature/connect/payment/trustly/ConnectTrustlyPaymentGraph.kt
@@ -1,0 +1,10 @@
+package com.hedvig.android.feature.connect.payment
+
+import androidx.navigation.NavGraphBuilder
+import com.hedvig.android.navigation.core.AppDestination
+import com.kiwi.navigationcompose.typed.composable
+
+fun NavGraphBuilder.connectTrustlyPaymentGraph() {
+  composable<AppDestination.ConnectPaymentTrustly>() {
+  }
+}
